Day 19
Beginning to create infrastructure for mapping Atom XML to Entities.

Created an Xml Extension method to make it easier to access simple properties.

Created a generic Factory: I'm not sure whether I like that, but for now I'm going to keep it.

Began implementing the conversion from Atom Entry XML to an Entry entity.