Massive restructuring, completion of unit tests, explicit naming of tables for certain relationships.
I decided it was getting to the point where I ought to split various portions of the app into its own projects, for potential reuse and for organizational purposes.

Also completed all of the unit tests that were reasonable to complete at this time.

Lastly, EFCodeFirst was doing well on the naming of most of the tables: I added explicit naming for those that it was mutilating.